+++ Vending machine.playground/xcuserdata/takahashishingo.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -4,7 +4,7 @@
 <dict>
 	<key>SchemeUserState</key>
 	<dict>
-		<key>Microwave (Playground).xcscheme</key>
+		<key>Vending machine (Playground).xcscheme</key>
 		<dict>
 			<key>isShown</key>
 			<false/>
